builds_complete/v0.6.30_beta-candidate-10_COMPLETE/requirements.txt
streamlit
pygame
python-chess
pyyaml
cairosvg
psutil
pillow
pandas
torch
numpy
dash
plotly
wmi
seaborn
firebase-admin>=6.2.0
firebase-functions
firebase-auth
google-cloud-storage>=2.9.0
google-cloud-firestore>=2.11.0
# Performance monitoring and profiling
matplotlib
nvidia-ml-py3  # Optional: for GPU monitoring
